% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBracketBetween.R
\name{gg_bracket_between}
\alias{gg_bracket_between}
\title{Draw bracket between samples}
\usage{
gg_bracket_between(data, sample_col, sample1, sample2, mean_col,
  group_col = NA, group1 = NA, group2 = NA, dodge.width = 0.9,
  extra_y_space = 0, ...)
}
\arguments{
\item{data}{The dataframe used in the plot}

\item{sample_col}{Column name of the samples in data}

\item{sample1}{The first sample to be compared}

\item{sample2}{The second sample to be compared}

\item{mean_col}{Column name of the dependent variable means. Usually the y-axis of the bargraph.}

\item{group_col}{Optional -- Column name of the sample groups if applicable.  Leave NA if there are no groups.}

\item{group1}{Optional -- The group to which the first sample belongs.  Leave NA if there are no groups.}

\item{group2}{Optional -- The group to which the second sample belongs.  Leave NA if there are no groups.}

\item{dodge.width}{Optional -- The width of position_dodge() in the ggplot if groups are present.  Usually 0.9.}

\item{extra_y_space}{Optional -- Numeric to adjust position of the bracket up or down from its default location.}

\item{...}{Additional optional parameters/asthetics.  e.g. color for font color or size for font size.}
}
\value{
ggplot2 annotation layer
}
\description{
ggplot2 layer to draw a bracket between two samples in a bargraph.  Used in conjunction with gg_ttest_between().
}
