% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTTestBetween.R
\name{t_test2}
\alias{t_test2}
\title{2-sample t-test}
\usage{
t_test2(mean1, mean2, sd1, sd2, n1, n2, m0 = 0, equal.variance = TRUE)
}
\arguments{
\item{mean1}{Mean for sample 1}

\item{mean2}{Mean for sample 2}

\item{sd1}{Standard deviation for sample 1}

\item{sd2}{Standard deviation for sample 2}

\item{n1}{Number of replicates for sample 1}

\item{n2}{Number of replicated for sample 2}

\item{m0}{Difference in means to test for.  Default is 0.}

\item{equal.variance}{Optional -- Boolean for whether variances are equal.  TRUE results in a student's t-test.  FALSE results in a Welch test.  Defaults to TRUE.}
}
\value{
Named list containing "Difference of means", "Std Error", "t", "df", and "p-value"
}
\description{
Function which performs a 2-sample t-test using sample statistics (mean, sd, n) instead of a lists of observations.  Not recommended for use outside gg_ttest_between().
}
